feat(odp-2): duplicate pagination component to prevent conflicts
